+++ resources/es/constants.js
@@ -25,3 +25,5 @@ export const DICTIONARIES = {
         url: 'https://freedict.org/',
     },
 };
+
+export const EASTER_EGG_DAYS = { };