+++ scripts/build-en-dictionary.mjs
@@ -10,6 +10,7 @@ import {
     BLOCKED_PARTS,
     LETTERS_NOT_ALLOWED_IN_WINNING_WORD,
     DICTIONARIES,
+    EASTER_EGG_DAYS,
 } from './../resources/en/constants.js';
 
 const LANG = 'en';
@@ -34,6 +35,7 @@ actionBuildDictionary(
         LETTERS_NOT_ALLOWED_IN_WINNING_WORD,
         DICTIONARIES,
         MAXIMUM_LENGHT_OF_ABOUT_LANGUAGE_WORD: 17,
+        EASTER_EGG_DAYS,
     },
     spellcheckerWords,
     winningWords,