+++ src/components/Panes/AboutLanguage/AboutLanguageLetters.tsx
@@ -8,7 +8,7 @@ import download from 'downloadjs';
 import { DictionaryInfo, DictionaryInfoLetters, ToastType } from '@common-types';
 
 import { useDispatch, useSelector } from '@store';
-import { setToast } from '@store/appSlice';
+import { track, setToast } from '@store/appSlice';
 
 import { getNow } from '@utils/date';
 import { capitalize } from '@utils/format';
@@ -24,6 +24,7 @@ import ButtonTile from '@components/Button/ButtonTile';
 import Modal from '@components/Modal/Modal';
 
 import KeyboardHeatmap from '@components/Charts/KeyboardHeatmap';
+import KeyboardLayoutPicker from '@components/Keyboard/KeyboardLayoutPicker';
 
 import AboutLanguageChartFooter from './AboutLanguageChartFooter';
 import AboutLanguageChartLanguageTitle from './AboutLanguageChartLanguageTitle';
@@ -63,6 +64,8 @@ const AboutLanguageLetters = ({
       const { stampOnlyTime } = getNow();
 
       download(dataUrl, `diffle-${gameLanguage}-${stampOnlyTime.replaceAll(':', '').replaceAll(' ', '')}.jpeg`);
+
+      dispatch(track({ name: `click_download_language_heatmap_${gameLanguage}` }));
     } catch (error) {
       // eslint-disable-next-line no-console
       console.error(error);
@@ -115,10 +118,11 @@ const AboutLanguageLetters = ({
               </div>
           </div>
           <div className="keyboard-heatmap-actions">
-              <Button className="keyboard-heatmap-action-edit" onClick={() => setIsOpen(true)} isInverted isText hasBorder={false}>
+              <Button onClick={() => setIsOpen(true)} isInverted isText hasBorder={false}>
                   <IconPencil />
               </Button>
-              <Button onClick={handleDownload} isInverted hasBorder={false}>
+              <KeyboardLayoutPicker shouldHideIfDisabled />
+              <Button className="about-language-download" onClick={handleDownload} isInverted hasBorder={false}>
                   <IconPicture />
                   <span>{t('common.download')}</span>
               </Button>