+++ CHANGELOG.md
@@ -11,7 +11,16 @@
 
 ---
 
-#### Version 3.18.1
+#### Version 3.19
+- **New**: Easter egg: sandbox live game mode added
+- **New**: Easter egg: custom emoji added
+- **New**: Copy only link added to share modal
+- **Fix**: The winning word cannot be a proper Roman numeral (as they often end up in dictionaries)
+- **New**: Keyboard layout picker added to the About language
+- **Fix**: Czech QWERTZ layout missing "y" restored
+- **New**: New GA events to track
+
+#### Version 3.18.1 (12.03.2024)
 - **Fix**: Align keyboard heatmaps when the title breaks the line
 - **Fix**: Fixed the incorrect keyboard layout for languages other than French in the About language